quickstart: remove assumption about flux binary relative path
some examples used `src/cmd/flux`, assuming the user was in the
flux-core source directory/repo.  This assumption no longer holds with
the flux-sched install instructions.

Remove usage of those relative paths and verbiage about flux adding
itself to the environment.  Reserve this verbiage for a future, separate
section in the docs.