coral: add instruction to load flux module
On LLNL-systems, it is technically OK to not load a `flux` module, since
we have a Flux installation (usually much older than the current
release) in `/usr/bin`, but on ORNL-systems, this does not work.  For
both systems, the situation is much better when the user loads our
provided `flux` module.