github actions: allow pip to upgrade required packages
Problem: The latest sphinx spelling package is incompatible with the
Sphinx package that comes with Debian 10 (which the action is running
in):

```
ERROR: sphinxcontrib-spelling 5.4.0 has requirement Sphinx>=3.0.0, but you'll have sphinx 2.4.4 which is incompatible.
```

The automatic pip install based on the `requirements.txt` does not pass
the `--upgrade` flag, and pip apparently doesn't support as fancy a dependency
resolution as something like Spack does.

Solution: Rather than mix OS system packages with the latest pip
packages, let's just go full bore with the bleeding edge from pip for
everything (it's almost like we are running Arch!).

The sphinx-action we are using doesn't have lots of flag (an intentional
decision by the maintainer), so just leverage the existing
`pre-build-command`.