adminguide: add warning about ranks vs hostnames
Problem: we are probably not going to be able to make 0.20.0
use hostnames in place of broker ranks in all situations
where it would be appropriate.

Add a 0.20.0 warning that this is a known usability issue.

Locate the warning in the resource exclusion section, an early
spot where it seems like hostnames might more convenient.