adminguide: sync fluke examples with the actual machine
The actual fluke [bootstrap] config proved to be more instructive
than the contrived examples in the admin guide.

Update the example to more closely match that actual config.

Sneak in an illustration of the idset->hostlist change in the
way the host array is parsed (second entry is now rank 1 rather
than rank 52).