coral: update test command for CUDA-enabled hwloc
Problem: Right now the documentation instructs the user to run `lstopo |
grep -i coproc`, which is an OK test, but it is possible:

1. for the right binary to be in the user's PATH but the wrong library
to be found by ld
2. the wrong binary to be in the user's PATH but the right library be
found by ld.

The latter is the case on Lassen where the module file only updates the
LD_LIBRARY_PATH.

Solution: Have the user test the intended behavior directly with
`flux start flux resource list`.

Closes #71