coral: switch the LLNL modulefile to load hwloc
Problem: users running on CORAL systems typically want to schedule and
run on GPUs "out-of-the-box", and our GPU-related instructions are all
the way at the bottom of the page.  We frequently get users asking why
they get an unsatisfiable request when asking for a GPU on Lassen.

Solution: adjust the modulefile on Lassen to auto-load the cuda-enabled
hwloc (the Summit modulefile already does this).  Tweak the docs to
note this, and adjust the paths that are `module use`'d to include the
`/usr/tce` directory that includes the `cuda` module.