coral: streamline load and launch instructions
Problem: the most recent installations of Flux on ORNL and LLNL CORAL
systems are configured with `--enable-pmix-bootstrap`, which greatly
simplifies the loading and launching of Flux on these systems.

Solution: streamline the instructions for the common case (i.e., using
the Flux team provided installation using this new PMIx option) and
demote the nitty-gritty details about the `pmi-shim` and `PMIX_*` env
vars to notes for anyone still leveraging them.