adminguide: drop manual install instructions
Problem: the admin guide gives instructions for manually installing
a Flux system instance, but this is unlikely to be an acceptable
practice at most sites and just decreases the signal to noise ratio
in the document.

Drop the installation description and add a 'note' box directing
would-be Flux package maintainers to open an issue for assistance.