index.rst: move coral and coral2 docs toward end of index
Problem: It seems strange to have CORAL and CORAL2 links come before
more general documentation such as how to run batch jobs in the
flux-docs index.

Move coral and coral2 links closer to the bottom of the index. We'll
keep the more generic docs near the top of the index.