adminguide: add tbon keepalive config info
Problem: TCP keepalives are enabled by default in flux-core 0.35.0,
but with the Linux default keepalive timing values, a powered off node
would not be detected for hours.

Add an example [tbon] config table that reduces detection time to 2.5m.
Reference flux-config-tbon(5) which explains the values and their
relationship to sysctl values in more detail.