adminguide: be more explicit about optional components
Problem: A new users found it confusing which components of
flux-frameowrk were optional when trying to build and configure a
Flux system instance.

Tag optional framework components with "(optional)" so it is very
clear from even a brief glance which components are not required.
Also, modify the parentheticals on flux-sched and flux-accounting
given the new optional tag on the project name.