adminguide: clarify that flux-core must be built with flux-security
Problem: If an admin is installing packages from source, it may not
be clear that flux-core must be configured with --with-flux-security
in order for a multi-user system instance to be functional.

Reorder the list of required framework packages to clarify the
order in which packages should be built and installed. Add a
note to the flux-core description that it needs to be built with
`--with-flux-security`.