faqs: remove link to flux-tree
Problem: The hyperlink to flux-tree no longer works on the FAQ page after
flux-framework/flux-sched#956, which moved flux-tree to the test suite and
changed paths in the repo. Its entry in the FAQ page is out of date and should
be updated to include more modern commands.

Remove the hyperlink to the flux-tree source code and instead make mention of
flux mini submit and flux mini bulksubmit. Add a hyperlink to the flux-mini
manual page.