quickstart: refer to LLNL site docs
Problem: the Quick Start document assumes the reader is exploring
Flux on their own but we now have users who must learn Flux because
it is their cluster's native resource manager.

Add a note box at the top of the Quick Start guide that refers to site
specific docs, specifically LLNL's, since that may be a better guide
for those users.