adminguide: make fluxion less optional
Problem: fluxion is listed as optional, but the expected primary use
case for multiple named queues is to partition resources, and at present,
sched-simple cannot handle that because it only processes the jobs in
FIFO priority order.  Also sched-simple doesn't do GPUs yet.

Drop the "optional" note.  Also drop instruction to install on management
node, as the scheduler is required on all nodes for batch jobs etc..