adminguide: add bootstrap.host parent directive
Problem: flux-core 0.45.0 now offers a way to configure a
system instance with a non-flat tree topology.

Mention the "parent" directive in a comment above the example
bootstrap.hosts defintion.  People can look up more info in the
flux-config-bootstrap(5) reference below it.  Smaller clusters
do not require it, so it's probably not appropriate to use in
the example.