adminguide: minor tune-up of overview architecture
Problem: the Overview architecture sub-section opens with a sentence
about the Flux broker executable, when what is trying to convey is
the notion of a Flux instance as a collection of brokers.  It seems
like an odd way to begin.

Reword that sentence to lead with the "flux instance" term.
Also reword another sentence in that section that seemed awkward.