docs: add glossary section
Problem: The documentation site contains terms, both Flux-specific and not,
that might be unclear to users who are reading the site and about Flux for
the first time.

Add a new Glossary section to the docs site, which will serve as a place where
terms, concepts, and acronyms are defined and explained in more detail than in
the places where they are introduced or mentioned for the first time.