expose flux comparison table
Problem: the comparison table is highly useful,
but hidden under a section of the FAQ and hard
to find.
Solution: add it as a navigation tab on the top
(the rationale that some will come to the docs
explicitly looking for it) along with as a link
in the first line of the quickstart.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>