glossary: add expanded definition for 'fair-share'
Problem: flux-accounting doesn't have a clear definition for fair-share
and its place in the priority plugin in terms of calculating a job's
priority.

Add 'fair-share' as a term in the Glossary section with an expanded
definition.

Make the "Glossary" section a referenceable one in the flux-accounting
guide so that when fair-share is mentioned in the Multi-Factor Priority
Plugin section, the expanded definition can be easily linked to.