contributing: add testsuite tips section
Problem: The Flux projects have very robust and powerful test
environments, but there isn't any documentation on how to leverage this
test environment for new developers or people who want to contribute to
the project for the first time.

Add a "Testing" section to the Contributing document on how to run
tests within a Flux project and combine it with other tools such as
valgrind.