project: flux-coral2, make search-able
The headers do not seem to be searchable, so for
a user to find the new link to flux-coral2 we need
to add a little bit of description that includes
the term. Since a user might also be looking
for a system that is relevant, I added the primary
and testing systems too.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>